Moved user registration endpoint to new Auth controller
