+++ passwordvault_api/passwordvault_domain/passwordvault_domain.csproj
@@ -10,6 +10,7 @@
       <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.3.0" />
       <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="9.0.0" />
       <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="9.0.0" />
+      <PackageReference Include="SendGrid" Version="9.29.3" />
     </ItemGroup>
 
 </Project>
